Agregar validaciones y mejorar manejo de errores en la API de elementos; actualizar esquemas de Swagger.

+++ src/pages/api/items/[id].ts
@@ -1,6 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import dbConnect from "@/lib/dbConnect";
 import Item from "@/schemas/Item";
+import mongoose from "mongoose";
 
 export default async function handler(
 	req: NextApiRequest,
@@ -34,32 +35,37 @@ export default async function handler(
 			 *             schema:
 			 *               $ref: '#/components/schemas/Item'
 			 *       400:
-			 *         description: Bad request. User ID must be an integer and larger than 0.
+			 *         description: Bad request.
 			 *         content:
 			 *           application/json:
 			 *             schema:
 			 *               $ref: '#/components/schemas/Error400'
 			 *       404:
-			 *         description: A user with the specified ID was not found.
+			 *         description: A item was not found.
 			 *         content:
 			 *           application/json:
 			 *            schema:
-       *              $ref: '#/components/schemas/Error404'
-			 *       5XX:
-			 *         description: Unexpected error.
+			 *              $ref: '#/components/schemas/Error404'
 			 */
 			try {
+				if (!mongoose.Types.ObjectId.isValid(id as string)) {
+					return res.status(400).json({
+						success: false,
+						error: "Bad Request",
+					});
+				}
 				const item = await Item.findById(id);
-
 				if (!item) {
-					return res.status(404).json({ success: false });
+					return res.status(404).json({
+						success: false,
+						error: "Data not Found",
+					});
 				}
 				res.status(200).json({ success: true, data: item });
 			} catch (error) {
-				// res.status(400).json({ success: false });
-				res.status(404).json({
+				res.status(500).json({
 					success: false,
-					error: "Data Not Found",
+					error: "Internal Server Error",
 				});
 			}
 			break;
@@ -91,27 +97,49 @@ export default async function handler(
 			 *           application/json:
 			 *             schema:
 			 *               $ref: '#/components/schemas/Item'
-			 *       404:
-			 *         description: Item not found
+			 *       400:
+			 *         description: A item was not found.
+			 *         content:
+			 *           application/json:
+			 *            schema:
+			 *              $ref: '#/components/schemas/Error400'
+			 *       500:
+			 *         description: Internal Server Error
+			 *         content:
+			 *           application/json:
+			 *            schema:
+			 *              $ref: '#/components/schemas/Error500'
 			 */
 			try {
+				if (!mongoose.Types.ObjectId.isValid(id as string)) {
+					return res.status(400).json({
+						success: false,
+						error: "Bad Request",
+					});
+				}
 				const item = await Item.findByIdAndUpdate(id, req.body, {
 					new: true,
 					runValidators: true,
 				});
 				if (!item) {
-					return res.status(404).json({ success: false });
+					return res.status(404).json({
+						success: false,
+						error: "Data not Found",
+					});
 				}
 				res.status(200).json({ success: true, data: item });
 			} catch (error) {
-				res.status(400).json({ success: false });
+				res.status(400).json({
+					success: false,
+					error: "Bad Request",
+				});
 			}
 			break;
 
 		case "DELETE":
 			/**
 			 * @swagger
-			 * /api/items/{id}:
+			 * /items/{id}:
 			 *   delete:
 			 *     summary: Delete an item
 			 *     description: Deletes an item from the database
@@ -123,20 +151,50 @@ export default async function handler(
 			 *         schema:
 			 *           type: string
 			 *     responses:
-			 *       204:
+			 *       200:
 			 *         description: Item deleted successfully
-			 *       404:
-			 *         description: Item not found
+			 *         content:
+			 *           application/json:
+			 *             schema:
+			 *               $ref: '#/components/schemas/Success200'
+			 *       400:
+			 *         description: A item was not found.
+			 *         content:
+			 *           application/json:
+			 *            schema:
+			 *              $ref: '#/components/schemas/Error400'
+			 *       500:
+			 *         description: Internal Server Error
+			 *         content:
+			 *           application/json:
+			 *            schema:
+			 *              $ref: '#/components/schemas/Error500'
 			 */
 			try {
-				const deletedItem = await Item.deleteOne({ _id: id });
-				console.log(deletedItem);
-				if (deletedItem.deletedCount === 0) {
-					return res.status(404).json({ success: false });
+				if (!mongoose.Types.ObjectId.isValid(id as string)) {
+					return res.status(400).json({
+						success: false,
+						error: "Bad Request",
+					});
 				}
-				res.status(204).json({ success: true, data: {} });
+				const deletedItem = await Item.deleteOne({ _id: id });
+				console.log({ deletedItem });
+				if (!deletedItem || deletedItem.deletedCount === 0) {
+					return res.status(404).json({
+						success: false,
+						error: "Data not Found",
+					});
+				}
+
+				res.status(200).json({
+					success: true,
+					data: { info: "Item deleted successfully" },
+				});
 			} catch (error) {
-				res.status(400).json({ success: false });
+				res.status(400).json({
+					success: false,
+					error: "Bad Request",
+				});
 			}
 			break;
